fix: complete all remaining type conversions for int fields
- booking_price_adjustment_service.dart: adj.amount.toDouble() for AdjustmentSummary
- sync_service.dart: convert all amount fields from toDouble() to toInt()
  - basicSalary
  - expenses.amount
  - cash_transactions.amount
  - payments.amount
  - hotel_day_ledger totalIncome/totalExpenses/pendingBalances

All database amount fields are now properly converted

+++ mobile/lib/services/booking_price_adjustment_service.dart
@@ -479,7 +479,7 @@ class BookingPriceAdjustmentService {
         adjustmentSummaries.add(AdjustmentSummary(
           uuid: adj.localUuid,
           type: type,
-          amount: adj.amount,
+          amount: adj.amount.toDouble(),  // ⭐ int to double
           effectiveHotelDay: adj.effectiveHotelDay,
           endHotelDay: adj.endHotelDay,
           nightsAffected: nightsAffected,
